Moved Dijkstra from "others" to "graphs"
